+++ helperFiles/machineLearning/modelControl/Models/pyTorch/emotionModelInterface/emotionModel/emotionModelHelpers/submodels/modelComponents/reversibleComponents/reversibleInterface.py
@@ -18,7 +18,7 @@ class reversibleInterface(nn.Module):
     def checkDualReconstruction(self, x1, x2, atol=1e-8, numLayers=10):
         # Initialize the forward data.
         reversibleInterface.changeDirections(False)
-        f1, f2 = x1.clone(), x2.clone().double()
+        f1, f2 = x1.clone().double(), x2.clone().double()
         t1 = time.time()
 
         # Perform the forward passes.
